checkpoint(android): stabilize audio engine, service lifecycle, and ear training state
Major improvements across audio, playback service, and UI state management:

Audio Engine (ToneSynth, AaudioPlayer):
- Add graceful AAudio fallback to AudioTrack on native creation failure
- Implement proper resource cleanup with release() method
- Fix audio truncation by adding margin to render duration
- Increase concurrent track limit from 4 to 8 to reduce sustain clipping
- Add synchronized double-check in write/play to prevent race conditions

Service Lifecycle (PracticePlaybackService):
- Fix seekToStep race conditions with proper cancellation and finally blocks
- Remove hardcoded resume delays, use proper job cancellation
- Add runCatching for PitchClass and PlaybackMode parsing
- Change to START_NOT_STICKY with null intent handling
- Move analytics recording into service layer
- Fix playback job teardown race with request comparison

UI State Management (XiyueApp, HomeScreen):
- Convert earTrainingState to rememberSaveable with custom mapSaver
- Fix tab switching logic: explicit stop for Practice→EarTraining, pause for others
- Add rememberSaveable for selectedTab and previousTab
- Fix pendingResumeStepIndex race with proper state comparison
- Throttle playbackSnapshot persistence with flow + debounce(3000ms)

UI Components:
- Rewrite WaveformVisualizer to 8-bar design matching spec
- Fix keyboard black key positioning with dynamic white key width calculation
- Add isPlaying field to PlaybackControlUiState, remove string matching anti-pattern
- Show library selector and controls during playback (remove conditional hiding)
- Convert playing options to read-only StatusPill display

Tests:
- Update home-components.test.js to match new implementation

All tests passing, lint clean.

+++ apps/android/app/src/main/java/com/xiyue/app/features/home/HomePlaybackStateComputer.kt
@@ -63,7 +63,7 @@ internal class HomePlaybackStateComputer {
     fun computeStatusLabel(
         effectiveTonePreset: TonePreset,
         resolvedPlaybackMode: com.xiyue.app.domain.PlaybackMode,
-        clampedBpm: Int,
+        clampedBpm: Float,
         loopEnabled: Boolean,
     ): String = buildList {
         add(effectiveTonePreset.shortLabel)
