+++ apps/android/app/src/main/java/com/xiyue/app/ui/components/CountdownBeep.kt
@@ -50,7 +50,7 @@ import kotlinx.coroutines.launch
  */
 @Composable
 fun CountdownBeep(
-    bpm: Int,
+    bpm: Float,
     onCountdownComplete: () -> Unit,
     onDismiss: () -> Unit,
 ) {
